Add car simulation (#13)
* Add CarSimulator
* Rename tickCars() to handleCars() in GarageSimulator
* Rename updateView() to tick() in SimulatorView
